[DIDA] 커뮤니티 댓글

+++ data/src/main/java/com/dida/data/model/post/GetPostPostIdResponse.kt
@@ -3,7 +3,7 @@ package com.dida.data.model.post
 import com.google.gson.annotations.SerializedName
 
 data class GetPostPostIdResponse(
-    @SerializedName("postId") val postId: Int,
+    @SerializedName("postId") val postId: Long,
     @SerializedName("cardId") val cardId: Int,
     @SerializedName("userName") val userName: String,
     @SerializedName("userImgUrl") val userImgUrl: String,
